# Faster TWAPs with smaller parts