@anxolin 2025-03-06
@MindyCoW @elena-zh I relaxed the TWAP restrictions too much to see how it behaves, and this will create some bad UXs.
I think sweet spot will be somewhere in the middle.
Can you help me find better reasonable numbers? I can create different versions with different values if you want to experiment.