@MindyCoW 2025-03-06
What exactly was the impetus for this? Is TWAP not fine the way it is?