@elena-zh 2025-03-06
> @MindyCoW @elena-zh I relaxed the TWAP restrictions too much to see how it behaves, and this will create some bad UXs. I think sweet spot will be somewhere in the middle. Can you help me find better reasonable numbers? I can create different versions with different values if you want to experiment.

Added my thoughts on this [here](https://cowservices.slack.com/archives/C0361CDG8GP/p1741258701303629?thread_ts=1741258042.459739&cid=C0361CDG8GP)